Reorganize trustregion submodules and update docstrings.
