+++ docs/source/refs.bib
@@ -827,4 +827,16 @@ year = {1975}
   pages={553-572}
 }
 
+@article{Bertsekas1982,
+author = {Bertsekas, Dimitri P.},
+title = {Projected Newton Methods for Optimization Problems with Simple Constraints},
+journal = {SIAM Journal on Control and Optimization},
+volume = {20},
+number = {2},
+pages = {221-246},
+year = {1982},
+doi = {10.1137/0320018},
+URL = {https://doi.org/10.1137/0320018},
+}
+
 @Comment{jabref-meta: databaseType:bibtex;}
